backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    LLM_API_URL = os.getenv("LLM_API_URL")
    LLM_API_KEY = os.getenv("LLM_API_KEY")
    LLM_MODEL = os.getenv("LLM_MODEL")
    EMBEDDING_MODEL = os.getenv("EMBEDDING_MODEL", "sentence-transformers/all-MiniLM-L6-v2")
    
    JIRA_URL = os.getenv("JIRA_URL")
    JIRA_USERNAME = os.getenv("JIRA_USERNAME")
    JIRA_TOKEN = os.getenv("JIRA_TOKEN")
    
    CONFLUENCE_URL = os.getenv("CONFLUENCE_URL")
    CONFLUENCE_USERNAME = os.getenv("CONFLUENCE_USERNAME")
    CONFLUENCE_TOKEN = os.getenv("CONFLUENCE_TOKEN")
    
    SIMULATED_REPO_PATH = os.path.abspath("simulated_repo_origin")
    WORKSPACES_DIR = os.path.abspath("workspaces")

config = Config()
