backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-huggingface
chromadb
atlassian-python-api
python-dotenv
pydantic
httpx
sentence-transformers
